Create the new Forecast Group component.
