+++ src/components/forecasts/CurrentForecast.vue
@@ -1,41 +1,19 @@
 <template>
   <v-container>
     <v-row>
-      <v-col>
-        <div class="text-center">
-          <v-pagination
-            v-model="page"
-            :length="pages"
-            :total-visible="7"
-          ></v-pagination></div
-      ></v-col>
+      <v-col>Current Forecasts</v-col>
     </v-row>
     <v-row>
-      <v-col
-        cols="4"
-        v-for="forecast in someForecasts"
-        :key="forecast.provider"
-      >
-        <WeatherForecastCard v-bind="forecast" />
-      </v-col>
+      <v-col><ForecastGroup v-bind:initialForecasts="forecasts"/></v-col>
     </v-row>
   </v-container>
 </template>
 
 <script>
-import WeatherForecastCard from "@/components/weather/WeatherForecastCard";
+import ForecastGroup from "@/components/weather/ForecastGroup";
 export default {
   name: "CurrentForecast",
-  components: { WeatherForecastCard },
-  computed: {
-    pages() {
-      const ceil = Math.ceil(this.forecasts.length / 3);
-      return ceil > 0 ? ceil : 1;
-    },
-    someForecasts: function() {
-      return this.forecasts.slice((this.page - 1) * 3, this.page * 3);
-    },
-  },
+  components: { ForecastGroup },
   data() {
     return {
       forecasts: [
@@ -148,7 +126,6 @@ export default {
           snow: 2,
         },
       ],
-      page: 1,
     };
   },
 };
